+++ src/main/java/com/minsoo/co/tireerpserver/model/request/tire/TireRequest.java
@@ -55,7 +55,7 @@ public class TireRequest {
 
     @Schema(name = "retail_price", description = "소비자금액")
     @Positive(message = "소비자금액은 양수여야 합니다.")
-    private Integer retailPrice;
+    private Long retailPrice;
 
     @Schema(name = "run_flat", description = "런플렛")
     private Boolean runFlat;
